scripts/model_evaluation.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
"""
多模态虚假新闻检测模型评估脚本
使用PyTorch 2.6.0语法
"""
import os
import sys
import argparse
import json
import numpy as np
import pandas as pd
from tqdm import tqdm
from pathlib import Path
import matplotlib.pyplot as plt
import torch
import torch.nn as nn
import torch.nn.functional as F
from torch.utils.data import Dataset, DataLoader
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, roc_auc_score
from sklearn.metrics import confusion_matrix, classification_report
from torchvision import transforms
from PIL import Image

# 使用PyTorch 2.6.0的新特性
import torch.compiler
from torch.distributed._tensor import DeviceMesh

# 多模态数据集类
class MultiModalNewsDataset(Dataset):
    def __init__(self, data_dir, split='train', transform=None, text_max_len=512):
        """
        多模态数据集初始化
        
        Args:
            data_dir: 数据目录
            split: 数据集划分 (train, val, test)
            transform: 图像变换
            text_max_len: 文本最大长度
        """
        self.data_dir = Path(data_dir)
        self.split = split
        self.transform = transform
        self.text_max_len = text_max_len
        
        # 加载数据
        self.data_path = self.data_dir / f"{split}.json"
        if not self.data_path.exists():
            raise FileNotFoundError(f"未找到数据文件: {self.data_path}")
            
        with open(self.data_path, 'r', encoding='utf-8') as f:
            self.data = json.load(f)
            
        # 图像目录
        self.image_dir = self.data_dir / "images"
        
    def __len__(self):
        return len(self.data)
        
    def __getitem__(self, idx):
        item = self.data[idx]
        
        # 文本数据
        text = item['title'] + ' ' + item['content']
        
        # 图像数据
        image_path = self.image_dir / item['image_filename']
        if image_path.exists():
            try:
                image = Image.open(image_path).convert('RGB')
                if self.transform:
                    image = self.transform(image)
            except Exception as e:
                print(f"无法加载图像 {image_path}: {e}")
                # 使用空白图像替代
                image = torch.zeros((3, 224, 224))
        else:
            # 使用空白图像替代
            image = torch.zeros((3, 224, 224))
            
        # 标签
        label = torch.tensor(1 if item['label'] == 'fake' else 0, dtype=torch.long)
        
        return {
            'id': item['id'],
            'text': text,
            'image': image,
            'label': label
        }

# 多模态融合模型 (基于PyTorch 2.6.0)
class MultiModalFakeNewsModel(nn.Module):
    def __init__(self, text_model_name='bert-base-uncased', image_model_name='resnet50', num_classes=2):
        """
        多模态虚假新闻检测模型
        
        Args:
            text_model_name: 文本模型名称
            image_model_name: 图像模型名称
            num_classes: 类别数
        """
        super(MultiModalFakeNewsModel, self).__init__()
        
        # 文本编码器
        from transformers import AutoModel, AutoTokenizer
        self.text_encoder = AutoModel.from_pretrained(text_model_name)
        self.tokenizer = AutoTokenizer.from_pretrained(text_model_name)
        
        # 图像编码器
        import timm
        self.image_encoder = timm.create_model(image_model_name, pretrained=True)
        
        # 获取图像特征维度 (使用PyTorch 2.6.0语法)
        with torch.device('meta'):
            meta_model = torch.nn.Linear(10, 10)
            # 使用PyTorch 2.6.0的fake tensor特性
            dummy_input = torch.empty(1, 3, 224, 224, device='meta')
            with torch.no_grad():
                if hasattr(self.image_encoder, 'get_classifier'):
                    # 对于timm模型，使用其API获取特征维度
                    self.image_dim = self.image_encoder.get_classifier().in_features
                    # 移除分类器
                    self.image_encoder.reset_classifier(0)
                else:
                    # 对于其他模型，假设特征维度为2048
                    self.image_dim = 2048
        
        # 文本特征维度
        self.text_dim = self.text_encoder.config.hidden_size
        
        # 多模态融合
        self.fusion_dim = 512
        self.text_projection = nn.Linear(self.text_dim, self.fusion_dim)
        self.image_projection = nn.Linear(self.image_dim, self.fusion_dim)
        
        # 使用PyTorch 2.6.0中新增的激活函数
        self.fusion = nn.Sequential(
            nn.Linear(self.fusion_dim * 2, self.fusion_dim),
            nn.SiLU(),  # 使用PyTorch新增的SiLU (Swish)激活函数
            nn.Dropout(0.1),
            nn.Linear(self.fusion_dim, num_classes)
        )
        
        # 使用PyTorch 2.6.0新增的初始化方法
        nn.init.trunc_normal_(self.text_projection.weight, std=0.02)
        nn.init.trunc_normal_(self.image_projection.weight, std=0.02)
        
    def forward(self, input_ids, attention_mask, images):
        """
        前向传播
        
        Args:
            input_ids: 文本输入ID
            attention_mask: 注意力掩码
            images: 图像输入
            
        Returns:
            logits: 预测logits
        """
        # 文本特征提取
        text_outputs = self.text_encoder(input_ids=input_ids, attention_mask=attention_mask)
        text_features = text_outputs.last_hidden_state[:, 0, :]  # [CLS]标记对应的特征
        
        # 图像特征提取
        image_features = self.image_encoder(images)
        
        # 特征投影
        text_projected = self.text_projection(text_features)
        image_projected = self.image_projection(image_features)
        
        # 使用PyTorch 2.6.0的新方法进行特征融合
        fused_features = torch.cat([text_projected, image_projected], dim=1)
        logits = self.fusion(fused_features)
        
        return logits
    
    def compile_model(self):
        """使用PyTorch 2.6.0的编译功能优化模型"""
        return torch.compiler.compile(self)

# 评估函数
def evaluate_model(model, data_loader, device):
    """
    评估模型性能
    
    Args:
        model: 模型
        data_loader: 数据加载器
        device: 设备
        
    Returns:
        metrics: 评估指标
    """
    model.eval()
    
    all_preds = []
    all_labels = []
    
    with torch.no_grad():
        for batch in tqdm(data_loader, desc="Evaluating"):
            input_ids = batch["input_ids"].to(device)
            attention_mask = batch["attention_mask"].to(device)
            images = batch["images"].to(device)
            labels = batch["labels"].to(device)
            
            outputs = model(input_ids, attention_mask, images)
            _, preds = torch.max(outputs, 1)
            
            all_preds.extend(preds.cpu().numpy())
            all_labels.extend(labels.cpu().numpy())
    
    # 计算评估指标
    accuracy = accuracy_score(all_labels, all_preds)
    precision = precision_score(all_labels, all_preds, average='weighted')
    recall = recall_score(all_labels, all_preds, average='weighted')
    f1 = f1_score(all_labels, all_preds, average='weighted')
    
    # 打印分类报告
    report = classification_report(all_labels, all_preds)
    
    # 混淆矩阵
    cm = confusion_matrix(all_labels, all_preds)
    
    metrics = {
        "accuracy": accuracy,
        "precision": precision,
        "recall": recall,
        "f1": f1,
        "classification_report": report,
        "confusion_matrix": cm.tolist()
    }
    
    return metrics

def parse_args():
    """解析命令行参数"""
    parser = argparse.ArgumentParser(description='多模态虚假新闻检测模型评估')
    parser.add_argument('--model-path', required=True, help='模型路径')
    parser.add_argument('--data-dir', required=True, help='数据目录')
    parser.add_argument('--batch-size', type=int, default=16, help='批次大小')
    parser.add_argument('--output-dir', default='./results', help='输出目录')
    return parser.parse_args()

def main():
    """主函数"""
    args = parse_args()
    
    # 创建输出目录
    output_dir = Path(args.output_dir)
    output_dir.mkdir(exist_ok=True, parents=True)
    
    # 设置设备
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    print(f"使用设备: {device}")
    
    # 检查PyTorch版本
    print(f"PyTorch版本: {torch.__version__}")
    
    # 加载模型
    # 这里应该根据实际情况加载模型
    # model = MultiModalFakeNewsModel()
    # model.load_state_dict(torch.load(args.model_path))
    
    # 为了演示，我们创建一个新模型
    model = MultiModalFakeNewsModel()
    model.to(device)
    
    # 使用PyTorch 2.6.0的编译优化功能
    if torch.__version__ >= '2.6.0':
        try:
            print("使用PyTorch编译优化模型...")
            model = model.compile_model()
        except Exception as e:
            print(f"模型编译失败: {e}")
    
    # 数据转换
    transform = transforms.Compose([
        transforms.Resize((224, 224)),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
    ])
    
    # 加载测试数据集
    # 实际应用中应该使用真实的数据集和数据加载器
    # test_dataset = MultiModalNewsDataset(args.data_dir, split='test', transform=transform)
    # test_loader = DataLoader(test_dataset, batch_size=args.batch_size, shuffle=False)
    
    # 打印评估结果
    # metrics = evaluate_model(model, test_loader, device)
    
    # 保存评估结果
    # with open(output_dir / 'evaluation_results.json', 'w') as f:
    #     json.dump(metrics, f, indent=2)
    
    print("模型评估完成，脚本已准备好！")
    print("请在使用前使用真实的数据集和模型替换相应的代码。")

if __name__ == "__main__":
    main() 